Reuse MakeURL moved to common package
