+++ lib/autoupdate/agent/config.go
@@ -19,7 +19,6 @@
 package agent
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io/fs"
@@ -30,6 +29,8 @@ import (
 	"github.com/google/renameio/v2"
 	"github.com/gravitational/trace"
 	"gopkg.in/yaml.v3"
+
+	"github.com/gravitational/teleport/lib/autoupdate"
 )
 
 const (
@@ -84,13 +85,13 @@ type Revision struct {
 	// Version is the version of Teleport.
 	Version string `yaml:"version" json:"version"`
 	// Flags describe the edition of Teleport.
-	Flags InstallFlags `yaml:"flags,flow,omitempty" json:"flags,omitempty"`
+	Flags autoupdate.InstallFlags `yaml:"flags,flow,omitempty" json:"flags,omitempty"`
 }
 
 // NewRevision create a Revision.
 // If version is not set, no flags are returned.
 // This ensures that all Revisions without versions are zero-valued.
-func NewRevision(version string, flags InstallFlags) Revision {
+func NewRevision(version string, flags autoupdate.InstallFlags) Revision {
 	if version != "" {
 		return Revision{
 			Version: version,
@@ -113,16 +114,16 @@ func NewRevisionFromDir(dir string) (Revision, error) {
 	}
 	switch flags := parts[1:]; len(flags) {
 	case 2:
-		if flags[1] != FlagFIPS.DirFlag() {
+		if flags[1] != autoupdate.FlagFIPS.DirFlag() {
 			break
 		}
-		out.Flags |= FlagFIPS
+		out.Flags |= autoupdate.FlagFIPS
 		fallthrough
 	case 1:
-		if flags[0] != FlagEnterprise.DirFlag() {
+		if flags[0] != autoupdate.FlagEnterprise.DirFlag() {
 			break
 		}
-		out.Flags |= FlagEnterprise
+		out.Flags |= autoupdate.FlagEnterprise
 		fallthrough
 	case 0:
 		return out, nil
@@ -135,11 +136,11 @@ func (r Revision) Dir() string {
 	// Do not change the order of these statements.
 	// Otherwise, installed versions will no longer match update.yaml.
 	var suffix string
-	if r.Flags&(FlagEnterprise|FlagFIPS) != 0 {
-		suffix += "_" + FlagEnterprise.DirFlag()
+	if r.Flags&(autoupdate.FlagEnterprise|autoupdate.FlagFIPS) != 0 {
+		suffix += "_" + autoupdate.FlagEnterprise.DirFlag()
 	}
-	if r.Flags&FlagFIPS != 0 {
-		suffix += "_" + FlagFIPS.DirFlag()
+	if r.Flags&autoupdate.FlagFIPS != 0 {
+		suffix += "_" + autoupdate.FlagFIPS.DirFlag()
 	}
 	return r.Version + suffix
 }
@@ -238,90 +239,4 @@ type FindResp struct {
 	InWindow bool `yaml:"in_window"`
 	// Jitter duration before an automated install
 	Jitter time.Duration `yaml:"jitter"`
-}
-
-// InstallFlags sets flags for the Teleport installation
-type InstallFlags int
-
-const (
-	// FlagEnterprise installs enterprise Teleport
-	FlagEnterprise InstallFlags = 1 << iota
-	// FlagFIPS installs FIPS Teleport
-	FlagFIPS
-)
-
-// NewInstallFlagsFromStrings returns InstallFlags given a slice of human-readable strings.
-func NewInstallFlagsFromStrings(s []string) InstallFlags {
-	var out InstallFlags
-	for _, f := range s {
-		for _, flag := range []InstallFlags{
-			FlagEnterprise,
-			FlagFIPS,
-		} {
-			if f == flag.String() {
-				out |= flag
-			}
-		}
-	}
-	return out
-}
-
-// Strings converts InstallFlags to a slice of human-readable strings.
-func (i InstallFlags) Strings() []string {
-	var out []string
-	for _, flag := range []InstallFlags{
-		FlagEnterprise,
-		FlagFIPS,
-	} {
-		if i&flag != 0 {
-			out = append(out, flag.String())
-		}
-	}
-	return out
-}
-
-// String returns the string representation of a single InstallFlag flag, or "Unknown".
-func (i InstallFlags) String() string {
-	switch i {
-	case 0:
-		return ""
-	case FlagEnterprise:
-		return "Enterprise"
-	case FlagFIPS:
-		return "FIPS"
-	}
-	return "Unknown"
-}
-
-// DirFlag returns the directory path representation of a single InstallFlag flag, or "unknown".
-func (i InstallFlags) DirFlag() string {
-	switch i {
-	case 0:
-		return ""
-	case FlagEnterprise:
-		return "ent"
-	case FlagFIPS:
-		return "fips"
-	}
-	return "unknown"
-}
-
-func (i InstallFlags) MarshalYAML() (any, error) {
-	return i.Strings(), nil
-}
-
-func (i InstallFlags) MarshalJSON() ([]byte, error) {
-	return json.Marshal(i.Strings())
-}
-
-func (i *InstallFlags) UnmarshalYAML(n *yaml.Node) error {
-	var s []string
-	if err := n.Decode(&s); err != nil {
-		return trace.Wrap(err)
-	}
-	if i == nil {
-		return trace.BadParameter("nil install flags while parsing YAML")
-	}
-	*i = NewInstallFlagsFromStrings(s)
-	return nil
 }
